Minor performance improvements
This change removes an unnecessary round trip to the Kubernetes cluster.
Instead of getting the $POD_NAME for the given $APP_NAME with
`kubectl get pods`, and then passing that through to `kubectl exec`,
we can pass the $APP_NAME directly to `kubectl exec`.

That saves a round trip to the server, making it slightly quicker to
connect overall.

I've also added `set -e` to stop the script immediately if something
goes wrong. And have redirected STDOUT to /dev/null when running
`aws eks update-kubeconfig` to avoid unnecessary noise in the console
output.