Use the default 'Assume Role TTL' value
This doesn't need to be specified. It's fine to stick with the default
of 60m.

The documentation I followed manually set it to 8h, but presumably
that's because it was intended for long-lived shell sessions. In our
case, we only need to make one authenticated call to AWS to run
`aws eks update-kubeconfig`. So it's fine for the credentials to timeout
any time after that call has been made.